refactor: Refactor the Connect Service module
