+++ src-tauri/src/service/mod.rs
@@ -1,8 +1,7 @@
 pub(crate) mod asset;
 pub(crate) mod auth_token;
 pub(crate) mod config;
-pub(crate) mod connect_methods;
-pub(crate) mod connect_token;
+pub(crate) mod connect;
 pub(crate) mod setting;
 pub(crate) mod user;
 pub(crate) mod version;
